@@ -5,11 +5,10 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^0.26.1",
-    "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.28.0",
-    "react-redux": "^7.2.6",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-hook-form": "^7.29.0",
+    "react-redux": "^7.2.7",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.0",
     "redux-thunk": "^2.4.1"
@@ -40,7 +39,7 @@
   },
   "devDependencies": {
     "redux": "^4.1.2",
-    "sass": "^1.49.9",
+    "sass": "^1.49.10",
     "sass-loader": "^12.6.0"
   }
 }
