@@ -1,4 +1,4 @@
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { StrictMode } from "react";
 import { Provider } from "react-redux";
 import store from "./store";
@@ -12,11 +12,12 @@ import "./styles/index.scss";
 store.dispatch(fetchUser());
 store.dispatch(fetchProduct());
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
   <StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </StrictMode>,
-  document.getElementById("root")
+  </StrictMode>
 );
